Add files via upload
The university management system is made using principles of Object Oriented Programming (OOP). This C++ program contains three classes i.e.; university café, gaming zone and teacher portal. This program allows the user to run any of three above classes pressing either 1, 2 or 3. The teacher portal prompts user to enter classes and subjects. The four classes are given and program asks user to select any of these after which different options are provided for respective assessments. The data was directly read from .csv files using file handling. Lastly all the assessments are assigned weights which then assign grades to each students. The program enters data on .csv file after generating grades.

 The café program displays menu and prices and provides different options to the user to select any of the items, add more items, and keep displaying the menu until user asks to display the bill and then display the bill and close the program.

The gaming zone program is basically a tic tac toe game that allows two users X and O to fill the boxes. The player who first fills the box with 3Xs or 3Os in a horizontal, vertical, or diagonal manner wins the game. In some cases, when none of the players succeeds in filling the  boxes horizontally, vertically, or diagonally with 3Xs or 3Os, then the game is considered a draw.

The entire project was done applying all the concepts and principles studied so far in CP and OOP which include use of classes, objects, inheritance, arrays, polymorphism, loops application and use of conditional statements.